# Improve performance by caching find_spec