--- a/astroid/manager.py
+++ b/astroid/manager.py
@@ -442,12 +442,11 @@ class AstroidManager:
         # pylint: disable=import-outside-toplevel
         from astroid.brain.helpers import register_all_brains
         from astroid.inference_tip import clear_inference_tip_cache
-        from astroid.interpreter._import.spec import clear_spec_cache
+        from astroid.interpreter._import.spec import _find_spec
         from astroid.interpreter.objectmodel import ObjectModel
         from astroid.nodes._base_nodes import LookupMixIn
         from astroid.nodes.scoped_nodes import ClassDef
 
-        clear_spec_cache()
         clear_inference_tip_cache()
         _invalidate_cache()  # inference context cache
 
@@ -461,6 +460,7 @@ class AstroidManager:
             util.is_namespace,
             ObjectModel.attributes,
             ClassDef._metaclass_lookup_attribute,
+            _find_spec,
         ):
             lru_cache.cache_clear()  # type: ignore[attr-defined]