--- a/tests/test_modutils.py
+++ b/tests/test_modutils.py
@@ -22,7 +22,6 @@ import astroid
 from astroid import modutils
 from astroid.const import PY310_PLUS
 from astroid.interpreter._import import spec
-from astroid.interpreter._import.spec import clear_spec_cache
 
 from . import resources
 
@@ -42,7 +41,7 @@ class ModuleFileTest(unittest.TestCase):
     package = "mypypa"
 
     def tearDown(self) -> None:
-        clear_spec_cache()
+        astroid.MANAGER.clear_cache()
         for k in list(sys.path_importer_cache):
             if "MyPyPa" in k:
                 del sys.path_importer_cache[k]