@crazybolillo 2024-04-05
I also obtained this call graph to find the best place to cache. The issue refers to `find_module` but I thought it was better to cache up the stack in `find_spec` because that method calls itself all the finders available. Since I thought it was cool and may be of interest, I am attaching it. You will need to zoom into it.

![output](https://github.com/pylint-dev/astroid/assets/40015596/6c14e929-9812-4770-982f-9bcad0b4a911)