@crazybolillo 2024-04-05
Also, I used the repository linked in the issue to obtain all the profiling data. I am still yet to profile the memory allocated for the process execution, will try to obtain that since I want to verify the hit an unbound cache may create.