@crazybolillo 2024-04-15
> Thank you for the analysis of what need to be cachec. Any reason not to use `functools.cache`or `functools.lru_cache` instead of using a global ?

I really wanted to use something standard, however the problem I found with functools caches is that they take into consideration all function arguments while I thought it only made sense to cache the requested module because:

- Search paths are lists so they need further processing to hash them
- I don't think search paths should be cached? If I already know where the package is find, does it matter if you give me different search paths? This point I am not so sure about, but it was my logic.